여태까지 작업한 거 올립니다.

+++ Assets/Scripts/Enemies/Boss/AngryDogAction.cs
@@ -38,13 +38,13 @@ public class AngryDogAction : EnemyAction {
                 temp = 1;
 
             //triple attack: shld check player's hallucinated 
-            player.ChangeHp( -(int) temp );
-            enemyItself.ChangeHp( (int) temp / 3 );
+            player.ChangeHp( -temp );
+            enemyItself.ChangeHp( temp / 3 );
             player.ChangeHp( -(int) temp );
             enemyItself.ChangeHp( (int) temp / 3 );
             //if(player.isHallucinated == true)
-            player.ChangeHp( -(int) temp );
-            enemyItself.ChangeHp( (int) temp / 3 );
+            player.ChangeHp( -temp );
+            enemyItself.ChangeHp( temp / 3 );
 
 
             if( player.Hp <= 0 )
